+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: "<No Username>" changed to: "Guest Account ([last 3 loginId characters])"
+
 ## 3.17.1 (2024-08-02)
 
 - fixed: DigitInput unknown returnKeyType for iOS